fix(cover) : Coveralls should not consider the helpers data which are useful for the tests.
