+++ pkg/helpers/db_seed.go
@@ -1,15 +1,12 @@
-package test
+package helpers
 
 import (
 	"database/sql"
-
-	"github.com/aerogear/mobile-security-service/pkg/models"
-
 	"github.com/sirupsen/logrus"
 )
 
 // Seed the database with some sample data
-func seedDatabase(db *sql.DB) {
+func SeedDatabase(db *sql.DB) {
 	_, err := db.Exec(`INSERT INTO app
 			(id, app_id, app_name, deleted_at)
 		VALUES 
@@ -37,59 +34,4 @@ func seedDatabase(db *sql.DB) {
 	if err != nil {
 		logrus.Println(err)
 	}
-}
-
-// GetMockAppList returns some dummy apps
-func GetMockAppList() []models.App {
-	apps := []models.App{
-		models.App{
-			ID:      "7f89ce49-a736-459e-9110-e52d049fc025",
-			AppID:   "com.aerogear.mobile_app_one",
-			AppName: "Mobile App One",
-		},
-		models.App{
-			ID:      "7f89ce49-a736-459e-9110-e52d049fc026",
-			AppID:   "com.aerogear.mobile_app_three",
-			AppName: "Mobile App Two",
-		},
-		models.App{
-			ID:      "7f89ce49-a736-459e-9110-e52d049fc027",
-			AppID:   "com.aerogear.mobile_app_three",
-			AppName: "Mobile App Three",
-		},
-	}
-	return apps
-}
-
-// GetMockAppVersionList returns some dummy app versions
-func GetMockAppVersionList() []models.Version {
-	versions := []models.Version{
-		models.Version{
-			ID:                   "55ebd387-9c68-4137-a367-a12025cc2cdb",
-			Version:              "1.0",
-			AppID:                "com.aerogear.mobile_app_one",
-			DisabledMessage:      "Please contact an administrator",
-			Disabled:             false,
-			NumOfCurrentInstalls: 1,
-			NumOfAppLaunches:     2,
-		},
-		models.Version{
-			ID:                   "59ebd387-9c68-4137-a367-a12025cc1cdb",
-			Version:              "1.1",
-			AppID:                "com.aerogear.mobile_app_one",
-			Disabled:             false,
-			NumOfCurrentInstalls: 0,
-			NumOfAppLaunches:     0,
-		},
-		models.Version{
-			ID:                   "59dbd387-9c68-4137-a367-a12025cc2cdb",
-			Version:              "1.0",
-			AppID:                "com.aerogear.mobile_app_two",
-			Disabled:             false,
-			NumOfCurrentInstalls: 0,
-			NumOfAppLaunches:     0,
-		},
-	}
-
-	return versions
 }
